resolve 3rd wave of PR feedback
