+++ zetaclient/chains/solana/signer/signer.go
@@ -7,7 +7,6 @@ import (
 	ethcommon "github.com/ethereum/go-ethereum/common"
 	"github.com/gagliardetto/solana-go"
 	"github.com/gagliardetto/solana-go/rpc"
-	"github.com/near/borsh-go"
 
 	"github.com/zeta-chain/zetacore/pkg/chains"
 	"github.com/zeta-chain/zetacore/pkg/coin"
@@ -69,102 +68,6 @@ func NewSigner(
 	}, nil
 }
 
-// SignMsgWithdraw signs a message for Solana gateway withdraw/withdraw_spl instruction
-func (signer *Signer) SignMsgWithdraw(
-	ctx context.Context,
-	params *types.OutboundParams,
-	height uint64,
-) (*contracts.MsgWithdraw, error) {
-	chain := signer.Chain()
-	// #nosec G115 always positive
-	chainID := uint64(signer.Chain().ChainId)
-	nonce := params.TssNonce
-	amount := params.Amount.Uint64()
-
-	// check receiver address
-	to, err := chains.DecodeSolanaWalletAddress(params.Receiver)
-	if err != nil {
-		return nil, errors.Wrapf(err, "cannot decode receiver address %s", params.Receiver)
-	}
-
-	// prepare withdraw msg and compute hash
-	msg := contracts.NewMsgWithdraw(chainID, nonce, amount, to)
-	msgHash := msg.Hash()
-
-	// sign the message with TSS to get an ECDSA signature.
-	// the produced signature is in the [R || S || V] format where V is 0 or 1.
-	signature, err := signer.TSS().Sign(ctx, msgHash[:], height, nonce, chain.ChainId, "")
-	if err != nil {
-		return nil, errors.Wrap(err, "Key-sign failed")
-	}
-	signer.Logger().Std.Info().Msgf("Key-sign succeed for chain %d nonce %d", chainID, nonce)
-
-	// attach the signature and return
-	return msg.SetSignature(signature), nil
-}
-
-// SignWithdrawTx signs the Solana gateway 'withdraw' transaction specified by 'msg'
-func (signer *Signer) SignWithdrawTx(ctx context.Context, msg contracts.MsgWithdraw) (*solana.Transaction, error) {
-	// create withdraw instruction with program call data
-	var err error
-	var inst solana.GenericInstruction
-	inst.DataBytes, err = borsh.Serialize(contracts.WithdrawInstructionParams{
-		Discriminator: contracts.DiscriminatorWithdraw(),
-		Amount:        msg.Amount(),
-		Signature:     msg.SigRS(),
-		RecoveryID:    msg.SigV(),
-		MessageHash:   msg.Hash(),
-		Nonce:         msg.Nonce(),
-	})
-	if err != nil {
-		return nil, errors.Wrap(err, "cannot serialize withdraw instruction")
-	}
-
-	// attach required accounts to the instruction
-	privkey := signer.solanaFeePayerKey
-	var accountSlice []*solana.AccountMeta
-	accountSlice = append(accountSlice, solana.Meta(privkey.PublicKey()).WRITE().SIGNER())
-	accountSlice = append(accountSlice, solana.Meta(signer.pda).WRITE())
-	accountSlice = append(accountSlice, solana.Meta(msg.To()).WRITE())
-	accountSlice = append(accountSlice, solana.Meta(signer.gatewayID))
-	inst.ProgID = signer.gatewayID
-	inst.AccountValues = accountSlice
-
-	// get a recent blockhash
-	recent, err := signer.client.GetRecentBlockhash(ctx, rpc.CommitmentFinalized)
-	if err != nil {
-		return nil, errors.Wrap(err, "GetRecentBlockhash error")
-	}
-
-	// create a transaction that wraps the instruction
-	tx, err := solana.NewTransaction(
-		[]solana.Instruction{
-			// TODO: outbound now uses 5K lamports as the fixed fee, we could explore priority fee and compute budget
-			// https://github.com/zeta-chain/node/issues/2599
-			// programs.ComputeBudgetSetComputeUnitLimit(computeUnitLimit),
-			// programs.ComputeBudgetSetComputeUnitPrice(computeUnitPrice),
-			&inst},
-		recent.Value.Blockhash,
-		solana.TransactionPayer(privkey.PublicKey()),
-	)
-	if err != nil {
-		return nil, errors.Wrap(err, "NewTransaction error")
-	}
-
-	// fee payer signs the transaction
-	_, err = tx.Sign(func(key solana.PublicKey) *solana.PrivateKey {
-		if key.Equals(privkey.PublicKey()) {
-			return &privkey
-		}
-		return nil
-	})
-	if err != nil {
-		return nil, errors.Wrap(err, "signer unable to sign transaction")
-	}
-
-	return tx, nil
-}
-
 // TryProcessOutbound - signer interface implementation
 // This function will attempt to build and sign a Solana transaction using the TSS signer.
 // It will then broadcast the signed transaction to the Solana chain.
@@ -195,15 +98,16 @@ func (signer *Signer) TryProcessOutbound(
 		Logger()
 
 	// support gas token only for Solana outbound
+	chainID := signer.Chain().ChainId
+	nonce := params.TssNonce
 	coinType := cctx.InboundParams.CoinType
 	if coinType != coin.CoinType_Gas {
-		logger.Error().Msgf("TryProcessOutbound: can only send SOL to the Solana network")
+		logger.Error().
+			Msgf("TryProcessOutbound: can only send SOL to the Solana network for chain %d nonce %d", chainID, nonce)
 		return
 	}
 
 	// sign gateway withdraw message by TSS
-	chainID := signer.Chain().ChainId
-	nonce := params.TssNonce
 	msg, err := signer.SignMsgWithdraw(ctx, params, height)
 	if err != nil {
 		logger.Error().Err(err).Msgf("TryProcessOutbound: SignMsgWithdraw error for chain %d nonce %d", chainID, nonce)
